John,
I made a few modifications to _form.html.erb, but I now have tag creation working smoothly. We need to decide whether or not to necessitate tags for creating events, if so, there will be more work to be done. I'm thinking that we can leave the tags as optional for now. Regardless, both scrape and user-created events now have tags automatically generated for them. Check it out!

Addison